requirements.txt
rumps>=0.4.0
broadlink>=0.18.0
python-dotenv>=1.0.0 